Add Linux:ZFS compatibility
The ZFS port on Linux behaves just like ext3 for certain tests.
This really has the most to do with Linux VFS and not the ZFS
backend implementation.  Since we expect ext3-like behavior
due to the VFS add a Linux:ZFS case to ensure these are not
marked as failures.